Consolidate DB base into database module
