+++ tests/test_alliance_bank_router.py
@@ -5,7 +5,7 @@ from fastapi import HTTPException
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from backend.db_base import Base
+from backend.database import Base
 from backend.models import Alliance, AllianceGrant, User
 from backend.routers.alliance_bank import GrantPayload, create_grant, list_grants
