+++ tests/test_alliance_management_router.py
@@ -7,7 +7,7 @@ import uuid
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from backend.db_base import Base
+from backend.database import Base
 from backend.models import Alliance, Kingdom, KingdomResources, User
 from backend.routers.alliance_management import (
     CreatePayload,
