+++ tests/test_alliance_members_router.py
@@ -7,7 +7,7 @@ from fastapi import HTTPException
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from backend.db_base import Base
+from backend.database import Base
 from backend.models import Alliance, AllianceMember, User
 from backend.routers.alliance_members import (
     RankPayload,
