+++ tests/test_alliance_policies_router.py
@@ -3,7 +3,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from fastapi import HTTPException
 
-from backend.db_base import Base
+from backend.database import Base
 from backend.models import Alliance, AlliancePolicy, User
 from backend.routers.alliance_policies import (
     PolicyPayload,
