+++ tests/test_alliance_votes_router.py
@@ -2,7 +2,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from fastapi import HTTPException
 
-from backend.db_base import Base
+from backend.database import Base
 from backend.models import Alliance, AllianceVote, AllianceVoteParticipant, User
 from backend.routers.alliance_votes import (
     BallotPayload,
