+++ tests/test_auth_router.py
@@ -4,7 +4,7 @@ import uuid
 from sqlalchemy import create_engine, text
 from sqlalchemy.orm import sessionmaker
 
-from backend.db_base import Base
+from backend.database import Base
 from backend.models import User
 from backend.routers import auth
 from backend.routers import forgot_password as fp
