+++ tests/test_battle_declare_endpoint.py
@@ -2,7 +2,7 @@ from fastapi import HTTPException
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from backend.db_base import Base
+from backend.database import Base
 from backend.models import Alliance, AllianceWar, User
 from backend.routers.battle import declare_alliance_battle, DeclarePayload
