+++ tests/test_battle_resolution_endpoint.py
@@ -2,7 +2,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
 from fastapi import HTTPException, status
-from backend.db_base import Base
+from backend.database import Base
 from backend.models import (
     BattleResolutionLog,
     CombatLog,
