+++ tests/test_battle_router.py
@@ -6,7 +6,7 @@ from fastapi import HTTPException
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from backend.db_base import Base
+from backend.database import Base
 from backend.models import WarScore
 from backend.routers.battle import get_battle_scoreboard
