+++ tests/test_black_market.py
@@ -7,7 +7,7 @@ import uuid
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from backend.db_base import Base
+from backend.database import Base
 from backend.models import BlackMarketListing
 from backend.routers.black_market import (
     BuyPayload,
