+++ tests/test_compose_router.py
@@ -3,7 +3,7 @@ import uuid
 from sqlalchemy import create_engine, text
 from sqlalchemy.orm import sessionmaker
 
-from backend.db_base import Base
+from backend.database import Base
 from backend.models import Alliance, AllianceNotice, PlayerMessage, User, War
 from backend.routers.compose import (
     MessagePayload,
