+++ tests/test_market_router.py
@@ -1,7 +1,7 @@
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from backend.db_base import Base
+from backend.database import Base
 from backend.models import MarketListing
 from backend.routers.market import (
     BuyPayload,
