+++ tests/test_noble_houses_router.py
@@ -5,7 +5,7 @@
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from backend.db_base import Base
+from backend.database import Base
 from backend.routers.noble_houses import (
     HousePayload,
     create_house,
