+++ tests/test_production_tick_service.py
@@ -2,7 +2,7 @@ import pytest
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from backend.db_base import Base
+from backend.database import Base
 from backend.models import KingdomResources, KingdomVillage, VillageProduction
 from services import production_tick_service
