+++ tests/test_reports_router.py
@@ -1,7 +1,7 @@
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from backend.db_base import Base
+from backend.database import Base
 from backend.models import UserReport
 from backend.routers.reports import ReportPayload, list_reports, submit_report
