+++ tests/test_resource_service.py
@@ -3,7 +3,7 @@ from fastapi import HTTPException
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from backend.db_base import Base
+from backend.database import Base
 from backend.models import KingdomResources, KingdomResourceTransfer
 from services import resource_service
