+++ tests/test_resources_router.py
@@ -5,7 +5,7 @@
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from backend.db_base import Base
+from backend.database import Base
 from backend.models import KingdomResources, User
 from backend.routers import resources
 from backend.routers.resources import get_resources
